services/semantic-python/semantic_server.py
"""
IRONCLAD Semantic Intent Classifier
Analyzes code diffs and classifies deployment intent.
"""

import os
from flask import Flask, jsonify, request
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

@app.route('/health', methods=['GET'])
def health():
    return jsonify({"status": "healthy", "service": "semantic-python"}), 200

@app.route('/api/v1/classify', methods=['POST'])
def classify_intent():
    """
    Classify the intent of a deployment diff.
    Stub implementation - full Claude integration coming.
    """
    data = request.get_json()
    return jsonify({
        "intent": "PENDING",
        "confidence": 0.0,
        "message": "Intent classifier not yet implemented"
    }), 200

if __name__ == '__main__':
    port = os.getenv('SEMANTIC_PORT', '8082')
    app.run(host='0.0.0.0', port=int(port), debug=True)
